requirements.txt
# Core UI Framework (Streamlit) 
# These packages handle the web interface and basic interactivity.
streamlit>=1.35.0
streamlit-authenticator>=0.3.2
streamlit-echarts>=0.4.0
audio-recorder-streamlit>=0.0.10

# AI & Language Models (Groq API) 
# Groq provides the high-speed inference for our transcript analysis.
groq>=0.9.0

# Speech-to-Text (STT) Engines 
# We support both Deepgram (cloud) and Whisper (local/OpenAI).
deepgram-sdk>=3.2.0
openai-whisper>=20231117

# Audio Handling & TTS 
# Tools for audio manipulation and text-to-speech generation.
pydub>=0.25.1
gTTS>=2.5.1
pyttsx3>=2.90

# Retrieval-Augmented Generation (RAG) Stack 
# LangChain orchestrates our AI flows, and Milvus stores our vector data.
langchain>=0.2.5
langchain-community>=0.2.5
langchain-text-splitters>=0.2.2
langchain-huggingface>=0.0.3
sentence-transformers>=3.0.0
pymilvus>=2.5.3

# Document Parsing 
# Support for reading PDF and Word documents for knowledge base ingestion.
pypdf>=4.2.0
python-docx>=1.1.0

# Data Processing & Visualization 
# Standard libraries for dataframes, math, and interactive charts.
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.22.0

# Reporting & Export 
# Utilities for generating PDF and Excel reports for quality audits.
reportlab>=4.2.2
openpyxl>=3.1.4

# Security & Utilities 
# Critical helpers for authentication, environment variables, and HTTP.
bcrypt>=4.1.3
PyYAML>=6.0.1
python-dotenv>=1.0.1
requests>=2.32.0
httpx>=0.27.0
